+++ app/src/main/AndroidManifest.xml
@@ -19,10 +19,13 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".activities.ArchivedListActivity"
-            android:parentActivityName=".activities.TaskListActivity"/>
         <activity
             android:name=".activities.TodoPagerActivity"
+            android:label="@string/app_name"
+            android:parentActivityName=".activities.TaskListActivity"
+            android:windowSoftInputMode="adjustResize|stateAlwaysVisible" />
+        <activity
+            android:name=".activities.ArchivedListActivity"
             android:label="@string/app_name"
             android:parentActivityName=".activities.TaskListActivity"
             android:windowSoftInputMode="adjustResize|stateAlwaysVisible" />
