+++ app/src/main/java/com/rahobbs/todo/activities/TaskListActivity.java
@@ -2,9 +2,7 @@ package com.rahobbs.todo.activities;
 
 import android.support.v4.app.Fragment;
 
-import com.rahobbs.todo.fragments.TasksListFragment;
-import com.rahobbs.todo.fragments.TodoListFragment;
-
+import com.rahobbs.todo.fragments.TaskListFragment;
 /**
  * Activity that hosts the to-do list fragment
  */
@@ -12,7 +10,7 @@ public class TaskListActivity extends SingleFragmentActivity {
 
     @Override
     protected Fragment createFragment() {
-        return new TasksListFragment();
+        return new TaskListFragment();
     }
 
 }
