+++ app/src/main/java/com/rahobbs/todo/fragments/ArchivedListFragment.java
@@ -1,6 +1,9 @@
 package com.rahobbs.todo.fragments;
 
+
+import android.content.Intent;
 import android.os.Bundle;
+import android.support.design.widget.FloatingActionButton;
 import android.support.v7.widget.LinearLayoutManager;
 import android.support.v7.widget.RecyclerView;
 import android.support.v7.widget.helper.ItemTouchHelper;
@@ -9,43 +12,51 @@ import android.view.View;
 import android.view.ViewGroup;
 
 import com.rahobbs.todo.R;
+import com.rahobbs.todo.activities.TodoPagerActivity;
 import com.rahobbs.todo.helpers.TodoItem;
 import com.rahobbs.todo.helpers.TodoLab;
-import com.rahobbs.todo.helpers.TodoListAdapter;
 import com.rahobbs.todo.interfaces.SimpleItemTouchHelperCallback;
 
 import java.util.Collections;
 import java.util.List;
 
 /**
- * Fragment to display list of archived tasks
+ * Created by rachael on 5/11/16.
  */
 public class ArchivedListFragment extends TodoListFragment {
-    private RecyclerView mTodoRecyclerView;
-    public ItemTouchHelper touchHelper;
-    public String type = "archived_list";
+
+    int listSize;
+    RecyclerView mTodoRecyclerView;
+
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setHasOptionsMenu(true);
+    }
 
     @Override
     public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
         View view = inflater.inflate(R.layout.archived_list_fragment, container, false);
         mTodoRecyclerView = (RecyclerView) view.findViewById(R.id.todo_recycler_view);
         mTodoRecyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
-        setFragmentType(type);
-
         updateUI();
+
         return view;
     }
 
     @Override
     public void onViewCreated(View view, Bundle savedInstanceState) {
         super.onViewCreated(view, savedInstanceState);
+
         RecyclerView recyclerView = (RecyclerView) view.findViewById(R.id.todo_recycler_view);
         recyclerView.setHasFixedSize(true);
         recyclerView.setLayoutManager(new LinearLayoutManager(getActivity()));
         TodoLab todoLab = TodoLab.get(getActivity());
         List<TodoItem> todoItems = todoLab.getArchivedItems();
+        listSize = todoItems.size();
         Collections.sort(todoItems, new TodoItem.PositionComparator());
-        TodoListAdapter adapter = new TodoListAdapter(todoItems, getContext(), getView(), getActivity(), this);
+        TodoListAdapter adapter = new TodoListAdapter(todoItems);
         recyclerView.setAdapter(adapter);
 
         ItemTouchHelper.Callback callback =
@@ -54,7 +65,6 @@ public class ArchivedListFragment extends TodoListFragment {
         touchHelper.attachToRecyclerView(mTodoRecyclerView);
     }
 
-    @Override
     public void updateUI() {
         TodoLab todoLab = TodoLab.get(getActivity());
         List<TodoItem> todoItems = todoLab.getArchivedItems();
@@ -62,9 +72,10 @@ public class ArchivedListFragment extends TodoListFragment {
 
         TodoListAdapter recyclerAdapter = (TodoListAdapter) mTodoRecyclerView.getAdapter();
         if (recyclerAdapter == null) {
-            mTodoRecyclerView.setAdapter(new TodoListAdapter(todoItems, getContext(), getView(), getActivity(), this));
+            mTodoRecyclerView.setAdapter(new TodoListAdapter(todoItems));
         } else {
             recyclerAdapter.setTodoItems(todoItems);
         }
     }
+
 }
