+++ app/src/main/res/menu/context_menu.xml
@@ -13,8 +13,14 @@
         android:icon="@drawable/ic_menu_delete"
         app:showAsAction="ifRoom"/>
     <item
+        android:id="@+id/share_list"
+        android:title="Share List"
+        app:showAsAction="ifRoom"
+        android:icon="@drawable/ic_share_list"/>
+
+    <item
         android:id="@+id/close_menu"
         android:title="Close"
         android:icon="@drawable/ic_close_menu"
-        app:showAsAction="ifRoom"/>
+        app:showAsAction="always"/>
 </menu>
